pkg/spanner/coveragedb: move package to pkg/coveragedb
