+++ tools/syz-covermerger/syz_covermerger.go
@@ -12,9 +12,9 @@ import (
 
 	"cloud.google.com/go/civil"
 	"github.com/google/syzkaller/dashboard/dashapi"
+	"github.com/google/syzkaller/pkg/coveragedb"
 	"github.com/google/syzkaller/pkg/covermerger"
 	"github.com/google/syzkaller/pkg/log"
-	"github.com/google/syzkaller/pkg/spanner/coveragedb"
 	_ "github.com/google/syzkaller/pkg/subsystem/lists"
 	"golang.org/x/exp/maps"
 )
